Allow simultaneous, asynchronous responses from players when attacked
Add a game customization option which toggles this behavior.

Simultaneous reactions are only allowed when each attacked player's
reaction does not depend on feedback from the attacker.

Simultaneous reactions are not allowed when the attacked players
must gain cards from the Supply, since if there are not enough
cards in the Supply then turn order becomes important.

Partially addresses #34, but still need to work on cards like Masquerade
which are not attacks but still require all players to react.