Minor bug fixes for Cornucopia expansion
Card class costs must be accessed directly via the `Card._cost` class attribute prior to game start.

In kingdoms where all available 2 $ and 3 $ cards are already in the supply and a Bane card must be chosen, one is swapped out of the kingdom to become the Bane card and is then replaced randomly.