Fix complex interaction between Tournament and Trader
The Tournament gains a Prize onto the player's deck, but if
the player has a Trader in their hand it will allow them
to return the Trader in exchange for a Silver. This was
throwing an exception since technically the Prize was never
in the Supply, so it needed to be returned to the Prizes pile.