157 redesign game log (#168)
The game log on the frontend has been totally redesigned into a right-hand sidebar.

The backend has been setup for future tree-ification of the game log but that code is not actually in use currently. The send and broadcast mechanisms have been slightly altered for compatibility with the new frontend.